requirements.txt
arithmetic==0.6.2
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
base58==2.0.1
base91==1.0.1
base92==1.0.3
BaseHash==3.0.8
bcrypt==3.2.0
beautifulsoup4==4.9.1
bjson==0.2.9
brainfuck==1.0.2
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cPython==0.0.6
crypto==1.4.1
cryptography==3.1
curl==0.0.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
Django==3.1.1
EasyProcess==0.3
entrypoint2==0.2.3
fire==0.3.1
Flask==1.1.2
gmpy2 @ file:///C:/Users/huanghong/Downloads/gmpy2-2.1.0a1-cp36-cp36m-win_amd64.whl
h5py==2.10.0
html5lib==1.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libnum==1.7.1
logger==1.4
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mpmath==1.1.0
Naked==0.1.31
numpy==1.19.2
opencv-python==4.2.0.34
pandas==1.1.2
paramiko==2.7.2
parso==0.7.1
pdfminer==20191125
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
pyasn1==0.4.5
pycparser==2.20
pycryptodome==3.9.8
pycryptodomex==3.9.8
Pygments==2.7.1
pylint==2.6.0
pymongo==3.11.0
PyMySQL==0.10.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyPdf==1.13
pypiwin32==223
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
PyYAML==5.3.1
pyzbar==0.1.8
qrcode==6.1
requests==2.24.0
scapy==2.4.4
scipy==1.5.2
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
sympy==1.6.2
termcolor==1.1.0
texttable==1.6.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
unrar==0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
zipp==3.4.0
zxing==0.11
